Moved back GetRenderPipelineMaterialList to HDUtils as it's HDRP specific
