+++ ScriptableRenderPipeline/HDRenderPipeline/HDUtils.cs
@@ -1,8 +1,37 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
 namespace UnityEngine.Experimental.Rendering.HDPipeline
 {
     public class HDUtils
     {
         public const RendererConfiguration k_RendererConfigurationBakedLighting = RendererConfiguration.PerObjectLightProbe | RendererConfiguration.PerObjectLightmaps | RendererConfiguration.PerObjectLightProbeProxyVolume;
+
+        public static List<RenderPipelineMaterial> GetRenderPipelineMaterialList()
+        {
+            var baseType = typeof(RenderPipelineMaterial);
+            var assembly = baseType.Assembly;
+
+            var types = assembly.GetTypes()
+                .Where(t => t.IsSubclassOf(baseType))
+                .Select(Activator.CreateInstance)
+                .Cast<RenderPipelineMaterial>()
+                .ToList();
+
+            // Note: If there is a need for an optimization in the future of this function, user can
+            // simply fill the materialList manually by commenting the code abode and returning a
+            // custom list of materials they use in their game.
+            //
+            // return new List<RenderPipelineMaterial>
+            // {
+            //    new Lit(),
+            //    new Unlit(),
+            //    ...
+            // };
+
+            return types;
+        }
 
         public static Matrix4x4 GetViewProjectionMatrix(Matrix4x4 worldToViewMatrix, Matrix4x4 projectionMatrix)
         {
