+++ TestbedPipelines/Fptl/SkyboxHelper.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
-using UnityEngine.Rendering;
 using UnityEngine;
-using UnityEngine.Experimental.Rendering;
 
 namespace UnityEngine.Experimental.Rendering
 {
